align with 10.0 test setup
