+++ system-tests/projects/runner-e2e-specs/cypress/e2e/sessions/multiple_sessions.cy.js
@@ -1,0 +1,10 @@
+it('t1', () => {
+  cy.session('user1', () => {
+    window.localStorage.foo = 'val'
+  })
+
+  cy.session('user2', () => {
+    window.localStorage.foo = 'val'
+    window.localStorage.bar = 'val'
+  })
+})
