+++ system-tests/projects/runner-e2e-specs/cypress/e2e/sessions/new_session.cy.js
@@ -1,0 +1,6 @@
+it('t1', () => {
+  const setupFn = cy.stub().as('runSetup')
+
+  cy.session('blank_session', setupFn)
+  cy.log('after')
+})
