+++ system-tests/projects/runner-e2e-specs/cypress/e2e/sessions/new_session_and_fails_validation.cy.js
@@ -1,0 +1,8 @@
+it('t1', () => {
+  const setupFn = cy.stub().as('runSetup')
+  const validateFn = cy.stub().returns(false).as('runValidation')
+
+  cy.session('blank_session', setupFn, {
+    validate: validateFn,
+  })
+})
