+++ system-tests/projects/runner-e2e-specs/cypress/e2e/sessions/new_session_with_validation.cy.js
@@ -1,0 +1,10 @@
+it('t1', () => {
+  const setupFn = cy.stub().as('runSetup')
+  const validateFn = cy.stub().as('runValidation')
+
+  cy.session('blank_session', setupFn, {
+    validate: validateFn,
+  })
+
+  cy.log('after')
+})
