+++ system-tests/projects/runner-e2e-specs/cypress/e2e/sessions/recreates_session_and_fails_validation.cy.js
@@ -1,0 +1,27 @@
+let setupFn
+let validateFn
+
+before(() => {
+  setupFn = cy.stub().as('runSetup')
+  validateFn = cy.stub().callsFake(() => {
+    if (validateFn.callCount >= 2) {
+      return false
+    }
+  }).as('runValidation')
+})
+
+it('t1', () => {
+  cy.session('user1', setupFn, {
+    validate: validateFn,
+  })
+
+  cy.log('after')
+})
+
+it('t2', () => {
+  cy.session('user1', setupFn, {
+    validate: validateFn,
+  })
+
+  cy.log('after')
+})
